Seperated test_executer.py from test_k8s.py
